% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankman_functions.R
\name{print.blankman_result}
\alias{print.blankman_result}
\title{Print Method for BlankmanF Result}
\usage{
\method{print}{blankman_result}(x, ...)
}
\arguments{
\item{x}{An object of class 'blankman_result'.}

\item{...}{Additional arguments (unused).}
}
\description{
Nicely prints the results of a blankman_result object.
}
