% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankman_functions.R
\name{run_friedman_test}
\alias{run_friedman_test}
\title{Run the Friedman Test}
\usage{
run_friedman_test(data, subject_col, treatment_col, response_col)
}
\arguments{
\item{data}{A data frame in long format.}

\item{subject_col}{Column name for the block/subject identifier.}

\item{treatment_col}{Column name for the treatment groups.}

\item{response_col}{Column name for the response variable.}
}
\value{
An object of class 'blankman_result' with test results and summary.
}
\description{
Conducts the Friedman test for repeated measures data using a tidy data frame.
}
